# perf(ops): Monomorphic sync op calls