--- a/bench_util/benches/op_baseline.rs
+++ b/bench_util/benches/op_baseline.rs
@@ -30,11 +30,11 @@ async fn op_pi_async() -> i64 {
 }
 
 fn bench_op_pi_json(b: &mut Bencher) {
-  bench_js_sync(b, r#"Deno.core.opSync("op_pi_json", null);"#, setup);
+  bench_js_sync(b, r#"Deno.core.ops.op_pi_json();"#, setup);
 }
 
 fn bench_op_nop(b: &mut Bencher) {
-  bench_js_sync(b, r#"Deno.core.opSync("op_nop", null, null, null);"#, setup);
+  bench_js_sync(b, r#"Deno.core.ops.op_nop();"#, setup);
 }
 
 fn bench_op_async(b: &mut Bencher) {