--- a/cli/tsc/99_main_compiler.js
+++ b/cli/tsc/99_main_compiler.js
@@ -15,6 +15,7 @@ delete Object.prototype.__proto__;
 ((window) => {
   /** @type {DenoCore} */
   const core = window.Deno.core;
+  const ops = core.ops;
 
   let logDebug = false;
   let logSource = "JS";
@@ -250,7 +251,7 @@ delete Object.prototype.__proto__;
       }
 
       this.#lastCheckTimeMs = timeMs;
-      return core.ops.op_is_cancelled();
+      return ops.op_is_cancelled();
     }
 
     throwIfCancellationRequested() {
@@ -274,11 +275,11 @@ delete Object.prototype.__proto__;
     fileExists(specifier) {
       debug(`host.fileExists("${specifier}")`);
       specifier = normalizedToOriginalMap.get(specifier) ?? specifier;
-      return core.unwrapOpResult(core.ops.op_exists({ specifier }));
+      return ops.op_exists({ specifier });
     },
     readFile(specifier) {
       debug(`host.readFile("${specifier}")`);
-      return core.unwrapOpResult(core.ops.op_load({ specifier })).data;
+      return ops.op_load({ specifier }).data;
     },
     getCancellationToken() {
       // createLanguageService will call this immediately and cache it
@@ -309,10 +310,8 @@ delete Object.prototype.__proto__;
       }
 
       /** @type {{ data: string; scriptKind: ts.ScriptKind; version: string; }} */
-      const { data, scriptKind, version } = core.unwrapOpResult(
-        core.ops.op_load(
-          { specifier },
-        ),
+      const { data, scriptKind, version } = ops.op_load(
+        { specifier },
       );
       assert(
         data != null,
@@ -339,13 +338,13 @@ delete Object.prototype.__proto__;
     },
     writeFile(fileName, data, _writeByteOrderMark, _onError, _sourceFiles) {
       debug(`host.writeFile("${fileName}")`);
-      return core.unwrapOpResult(core.ops.op_emit(
+      return ops.op_emit(
         { fileName, data },
-      ));
+      );
     },
     getCurrentDirectory() {
       debug(`host.getCurrentDirectory()`);
-      return cwd ?? core.unwrapOpResult(core.ops.op_cwd());
+      return cwd ?? ops.op_cwd();
     },
     getCanonicalFileName(fileName) {
       return fileName;
@@ -361,10 +360,10 @@ delete Object.prototype.__proto__;
       debug(`  base: ${base}`);
       debug(`  specifiers: ${specifiers.join(", ")}`);
       /** @type {Array<[string, ts.Extension] | undefined>} */
-      const resolved = core.unwrapOpResult(core.ops.op_resolve({
+      const resolved = ops.op_resolve({
         specifiers,
         base,
-      }));
+      });
       if (resolved) {
         const result = resolved.map((item) => {
           if (item) {
@@ -384,7 +383,7 @@ delete Object.prototype.__proto__;
       }
     },
     createHash(data) {
-      return core.unwrapOpResult(core.ops.op_create_hash({ data })).hash;
+      return ops.op_create_hash({ data }).hash;
     },
 
     // LanguageServiceHost
@@ -399,7 +398,7 @@ delete Object.prototype.__proto__;
       if (scriptFileNamesCache) {
         return scriptFileNamesCache;
       }
-      return scriptFileNamesCache = core.ops.op_script_names();
+      return scriptFileNamesCache = ops.op_script_names();
     },
     getScriptVersion(specifier) {
       debug(`host.getScriptVersion("${specifier}")`);
@@ -412,9 +411,7 @@ delete Object.prototype.__proto__;
       if (scriptVersionCache.has(specifier)) {
         return scriptVersionCache.get(specifier);
       }
-      const scriptVersion = core.unwrapOpResult(
-        core.ops.op_script_version({ specifier }),
-      );
+      const scriptVersion = ops.op_script_version({ specifier });
       scriptVersionCache.set(specifier, scriptVersion);
       return scriptVersion;
     },
@@ -435,9 +432,9 @@ delete Object.prototype.__proto__;
         };
       }
 
-      const fileInfo = core.unwrapOpResult(core.ops.op_load(
+      const fileInfo = ops.op_load(
         { specifier },
-      ));
+      );
       if (fileInfo) {
         scriptVersionCache.set(specifier, fileInfo.version);
         return ts.ScriptSnapshot.fromString(fileInfo.data);
@@ -568,10 +565,10 @@ delete Object.prototype.__proto__;
 
     performanceProgram({ program });
 
-    core.unwrapOpResult(core.ops.op_respond({
+    ops.op_respond({
       diagnostics: fromTypeScriptDiagnostic(diagnostics),
       stats: performanceEnd(),
-    }));
+    });
     debug("<<< exec stop");
   }
 
@@ -580,7 +577,7 @@ delete Object.prototype.__proto__;
    * @param {any} data
    */
   function respond(id, data = null) {
-    core.unwrapOpResult(core.ops.op_respond({ id, data }));
+    ops.op_respond({ id, data });
   }
 
   /**
@@ -943,7 +940,7 @@ delete Object.prototype.__proto__;
   // ensure the snapshot is setup properly.
   /** @type {{ buildSpecifier: string; libs: string[] }} */
 
-  const { buildSpecifier, libs } = core.ops.op_build_info();
+  const { buildSpecifier, libs } = ops.op_build_info();
   for (const lib of libs) {
     const specifier = `lib.${lib}.d.ts`;
     // we are using internal APIs here to "inject" our custom libraries into