--- a/core/01_core.js
+++ b/core/01_core.js
@@ -187,14 +187,14 @@
     if (!hasPromise(promiseId)) {
       return;
     }
-    unwrapOpResult(ops.op_ref_op(promiseId));
+    ops.op_ref_op(promiseId);
   }
 
   function unrefOp(promiseId) {
     if (!hasPromise(promiseId)) {
       return;
     }
-    unwrapOpResult(ops.op_unref_op(promiseId));
+    ops.op_unref_op(promiseId);
   }
 
   function resources() {
@@ -211,7 +211,7 @@
   }
 
   function queueMicrotask(cb) {
-    return unwrapOpResult(ops.op_queue_microtask(cb));
+    return ops.op_queue_microtask(cb);
   }
 
   // Some "extensions" rely on "BadResource" and "Interrupted" errors in the
@@ -235,7 +235,6 @@
 
   // Extra Deno.core.* exports
   const core = ObjectAssign(globalThis.Deno.core, {
-    unwrapOpResult,
     opAsync,
     opSync,
     resources,
@@ -253,45 +252,44 @@
     opCallTraces,
     refOp,
     unrefOp,
-    close: (rid) => unwrapOpResult(ops.op_close(rid)),
-    tryClose: (rid) => unwrapOpResult(ops.op_try_close(rid)),
+    close: (rid) => ops.op_close(rid),
+    tryClose: (rid) => ops.op_try_close(rid),
     read: opAsync.bind(null, "op_read"),
     write: opAsync.bind(null, "op_write"),
     shutdown: opAsync.bind(null, "op_shutdown"),
-    print: (msg, isErr) => unwrapOpResult(ops.op_print(msg, isErr)),
-    setMacrotaskCallback: (fn) =>
-      unwrapOpResult(ops.op_set_macrotask_callback(fn)),
-    setNextTickCallback: (fn) =>
-      unwrapOpResult(ops.op_set_next_tick_callback(fn)),
+    print: (msg, isErr) => ops.op_print(msg, isErr),
+    setMacrotaskCallback: (fn) => ops.op_set_macrotask_callback(fn),
+    setNextTickCallback: (fn) => ops.op_set_next_tick_callback(fn),
     runMicrotasks: () => ops.op_run_microtasks(),
     hasTickScheduled: () => ops.op_has_tick_scheduled(),
-    setHasTickScheduled: (bool) =>
-      unwrapOpResult(ops.op_set_has_tick_scheduled(bool)),
-    evalContext: (source, specifier) =>
-      unwrapOpResult(ops.op_eval_context(source, specifier)),
+    setHasTickScheduled: (bool) => ops.op_set_has_tick_scheduled(bool),
+    evalContext: (
+      source,
+      specifier,
+    ) => ops.op_eval_context(source, specifier),
     createHostObject: () => ops.op_create_host_object(),
-    encode: (text) => unwrapOpResult(ops.op_encode(text)),
-    decode: (buffer) => unwrapOpResult(ops.op_decode(buffer)),
-    serialize: (value, options, errorCallback) =>
-      unwrapOpResult(ops.op_serialize(value, options, errorCallback)),
-    deserialize: (buffer, options) =>
-      unwrapOpResult(ops.op_deserialize(buffer, options)),
-    getPromiseDetails: (promise) =>
-      unwrapOpResult(ops.op_get_promise_details(promise)),
+    encode: (text) => ops.op_encode(text),
+    decode: (buffer) => ops.op_decode(buffer),
+    serialize: (
+      value,
+      options,
+      errorCallback,
+    ) => ops.op_serialize(value, options, errorCallback),
+    deserialize: (buffer, options) => ops.op_deserialize(buffer, options),
+    getPromiseDetails: (promise) => ops.op_get_promise_details(promise),
     getProxyDetails: (proxy) => ops.op_get_proxy_details(proxy),
     isProxy: (value) => ops.op_is_proxy(value),
     memoryUsage: () => ops.op_memory_usage(),
-    setWasmStreamingCallback: (fn) =>
-      unwrapOpResult(ops.op_set_wasm_streaming_callback(fn)),
-    abortWasmStreaming: (rid, error) =>
-      unwrapOpResult(ops.op_abort_wasm_streaming(rid, error)),
-    destructureError: (error) =>
-      unwrapOpResult(ops.op_destructure_error(error)),
+    setWasmStreamingCallback: (fn) => ops.op_set_wasm_streaming_callback(fn),
+    abortWasmStreaming: (
+      rid,
+      error,
+    ) => ops.op_abort_wasm_streaming(rid, error),
+    destructureError: (error) => ops.op_destructure_error(error),
     terminate: (exception) => ops.op_terminate(exception),
     opNames: () => ops.op_op_names(),
     eventLoopHasMoreWork: () => ops.op_event_loop_has_more_work(),
-    setPromiseRejectCallback: (fn) =>
-      unwrapOpResult(ops.op_set_promise_reject_callback(fn)),
+    setPromiseRejectCallback: (fn) => ops.op_set_promise_reject_callback(fn),
   });
 
   ObjectAssign(globalThis.__bootstrap, { core });