@Phosra 2022-07-29
IMO, It's reasonable to just do a swift inlining pass that ignores eluding `unwrapOpResult`, if it makes it easier to get a first iteration done here.