@aapoalas 2022-07-29
> If it makes it easier to get a first iteration done here, IMO it's reasonable to just do a swift inlining pass that ignores eluding `unwrapOpResult`.

Yeah, I think a lot of the files I'll just be doing that.