@aapoalas 2022-08-11
So, apparently this is ready for review and merge if nothing out of order is found. The tests failed once in parsing strace output but the failure did not repeat on another go around, so apparently it was a flake.